requirements.txt
opensearch-py
requests
pandas
kafka-python-ng
sentence-transformers
fastapi
uvicorn[standard]

